fix(defaults): set clientlibs defaults to an empty list
BREAKING CHANGE: The golang clientib is not required for
the regular use of this formula but, if you already expected it to be
installed by default, you'll need to update your pillars to do so.
Running this version of the formula over previous ones won't break the
minions, only skip clientlibs.